Add formatting test and move parsing test into correct file
